+++ be/src/storage/rowset/rowset_meta.h
@@ -199,7 +199,7 @@ public:
         if (!is_segments_overlapping()) {
             score = 1;
         } else {
-            score = num_segments();
+            score = static_cast<uint32_t>(num_segments());
             CHECK(score > 0);
         }
         return score;
@@ -225,7 +225,7 @@ public:
 
 private:
     bool _deserialize_from_pb(std::string_view value) {
-        return _rowset_meta_pb->ParseFromArray(value.data(), value.size());
+        return _rowset_meta_pb->ParseFromArray(value.data(), static_cast<int>(value.size()));
     }
 
     void _init() {
