+++ be/src/util/priority_thread_pool.hpp
@@ -146,7 +146,7 @@ public:
     }
 
     void set_num_thread(int num_thread) {
-        size_t num_thread_in_pool = _threads.size();
+        int num_thread_in_pool = static_cast<int>(_threads.size());
         if (num_thread > num_thread_in_pool) {
             increase_thr(num_thread - num_thread_in_pool);
         } else if (num_thread < num_thread_in_pool) {
